Dev/0.9.5 (#850)
* muse spark, test speed

* gpt-5.6 class, meta, tidyup

* commit a2a examples

* example update

---------

+++ src/fast_agent/acp/terminal_runtime.py
@@ -251,7 +251,7 @@ class ACPTerminalRuntime:
         session_id: str,
         activation_reason: str,
         logger_instance=None,
-        timeout_seconds: int = 90,
+        timeout_seconds: float = 90,
         tool_handler: "ToolExecutionHandler | None" = None,
         default_output_byte_limit: int = DEFAULT_TERMINAL_OUTPUT_BYTE_LIMIT,
         permission_handler: "ToolPermissionHandler | None" = None,
